Compute variance inflation factor from uncertainty in responding sample size (for integer \(n_h\))

Computes \(\zeta_h(n_h, \bar{\phi}_h):=\mathrm{E}(r_h)\mathrm{E}\left(\frac{1}{r_h}\right)\), as defined in the paper. If there are any strata where the allocation may lead to \(\mathrm{E}(r_h) < r_h^{LB}\) for user-specified \(r_h^{LB}\) (3.5 by default), then \(\zeta_h(.)\) is evaluated using \(n_h':= max\left(n_h, \left\lceil \frac{r_h^{LB}}{\bar{\phi}_h}\right\rceil\right)\) in place of \(n_h\). Continuous values of \(n_h\) are rounded (by default).

Usage

calc_zeta_discrete(
  n_h,
  phibar_h,
  rh_min = 3.5,
  round_flag = TRUE,
  verbose_flag = FALSE
)

Arguments

n_h

(vector) strata sample sizes before nonresponse (\(n_h\))

phibar_h

(vector) strata response propensities (\(\bar{\phi}_h\))

rh_min

(scalar) minimum target respondents per stratum (\(r_h^{LB}\)); default is 3.5

round_flag

(bool) specifies whether to round continuous allocations (versus throwing an error)

verbose_flag

(bool) flag on whether to provide detailed results

Value

vector of length \(H\) containing \(\left\{\zeta_h(n_h', \bar{\phi}_h):h=1,2,...,H\right\}\), where \(n_h'\) is the larger of \(n_h\) or \(\frac{r_h^{LB}}{\bar{\phi}_h}\)

Details

In our paper, we assumed that the number of respondents in stratum \(h\) can be modeled as standard binomial with support for zero removed (i.e., zero-truncated binomial; see dtruncbinom()), written as \(r_h \sim TBinom(n_h, \bar{\phi}_h)\), where \(n_h\) is the number of invitees in stratum \(h\), \(\bar{\phi}_h\) is the average response propensity within stratum \(h\), and where the unit-level response propensities are assumed constant within strata. Our paper defines the function \(\zeta_h(n_h, \bar{\phi}_h):=\mathrm{E}(r_h)\mathrm{E}\left(\frac{1}{r_h}\right)\). This quantity is a variance inflation term that captures the effect of variability in the number of respondents for a given allocation.

The current function (calc_zeta_discrete()) computes \(\zeta_h(n_h', \bar{\phi}_h)\), where we use \(n_h':= max\left(n_h, \left\lceil \frac{r_h^{LB}}{\bar{\phi}_h}\right\rceil\right)\) to avoid underallocating to strata with too few expected respondents, and where \(r_h^{LB}\) is some given lower bound on the number of expected respondents. By default, we set \(r_h^{LB} = 3.5\), since the truncated binomial distribution may sometimes be a poor approximation for the binomial distribution below these levels, and as we observed numerically that \(\zeta_h(n_h, \bar{\phi}_h)\) is roughly maximized for various \(\bar{\phi}_h\) (fixed at levels between .01 and 1) when \(n_h \approx \frac{3.5}{\bar{\phi}_h}\).

Function is only defined for discrete \(n_h\); input round_flag controls whether \(n_h\) is rounded to the nearest integer (as opposed to throwing an error).

See also

Examples

if (FALSE) {
#Basic example
calc_zeta_discrete(n_h = c(100, 200, 300, 300),
                  phibar_h = c(.03, .02, .05, .001))

#Verbose info shows that strata 1,5 have their sample sizes increased
#   due to having fewer than 3.5 expected respondents
calc_zeta_discrete(n_h = c(100, 200, 300, 300),
                   phibar_h = c(.03, .02, .05, .001),
                   verbose_flag = TRUE)

#Similar to above, but removes the minimum number of respondents,
# which changes the values for strata 1 and 4.
calc_zeta_discrete(n_h = c(100, 200, 300, 300),
                  phibar_h = c(.03, .02, .05, .001),
                  rh_min = 0,
                  verbose_flag = TRUE)
}